+++ src/components/MapView/index.tsx
@@ -5,7 +5,7 @@ import { createStyles, WithStyles, withStyles } from '@material-ui/core';
 
 import Boundaries from './Boundaries';
 import NSOLayers from './NSOLayers';
-import Layers from './Layers';
+import WMSLayers from './WMSLayers';
 import Legends from './Legends';
 import DateSelector from './DateSelector';
 import { dateRangeSelector, layersSelector } from '../../context/mapStateSlice';
@@ -23,7 +23,7 @@ function MapView({ classes }: MapViewProps) {
   const serverAvailableDates = useSelector(availableDatesSelector);
 
   const baselineLayers = layers.filter(layer => layer.type === 'nso');
-  const serverLayers = layers.filter(layer => layer.type in ['wms']);
+  const serverLayers = layers.filter(layer => layer.type === 'wms');
 
   const selectedLayerDates = layers
     .map(({ serverLayer }) =>
@@ -51,7 +51,7 @@ function MapView({ classes }: MapViewProps) {
       >
         <Boundaries />
         <NSOLayers layers={baselineLayers} />
-        <Layers layers={serverLayers} selectedDate={startDate} />
+        <WMSLayers layers={serverLayers} selectedDate={startDate} />
       </MapboxMap>
       <DateSelector availableDates={selectedLayerDates} />
       <Legends layers={layers} />
